+++ frontend/src/components/Shop/AllProducts.jsx
@@ -7,20 +7,37 @@ import { Link } from "react-router-dom";
 import { getAllProductsShop } from "../../redux/actions/product";
 import { deleteProduct } from "../../redux/actions/product";
 import Loader from "../Layout/Loader";
+import { toast } from "react-toastify";
 
 const AllProducts = () => {
-  const { products, isLoading } = useSelector((state) => state.products);
+  const { products, isLoading, error, message } = useSelector((state) => state.products);
   const { seller } = useSelector((state) => state.seller);
 
   const dispatch = useDispatch();
 
   useEffect(() => {
     dispatch(getAllProductsShop(seller._id));
-  }, [dispatch]);
+  }, [dispatch, seller._id]);
 
-  const handleDelete = (id) => {
-    dispatch(deleteProduct(id));
-    window.location.reload();
+  useEffect(() => {
+    if (error) {
+      toast.error(error);
+      dispatch({ type: "clearErrors" });
+    }
+
+  }, [dispatch, error, seller._id]);
+
+  const handleDelete = async (id) => {
+    if (window.confirm("Are you sure you want to delete this product?")) {
+      try {
+        await dispatch(deleteProduct(id));
+        toast.success("Product deleted successfully!");
+        dispatch(getAllProductsShop(seller._id));
+        dispatch({ type: "clearErrors" });
+      } catch (error) {
+        toast.error("Failed to delete product");
+      }
+    }
   };
 
   const columns = [
